# services/identity-service/src/main/java/com/cafeadmin/identity/api/AuthController.java
package com.cafeadmin.identity.api;

import java.util.Map;

import org.springframework.http.HttpHeaders;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.cafeadmin.identity.api.dto.LoginRequest;
import com.cafeadmin.identity.api.dto.UserResponse;
import com.cafeadmin.identity.service.AuthService;

import jakarta.validation.Valid;

@RestController
@RequestMapping("/api/auth")
public class AuthController {

    private final AuthService authService;

    public AuthController(AuthService authService) {
        this.authService = authService;
    }

    @PostMapping("/login")
    public Map<String, Object> login(@Valid @RequestBody LoginRequest request) {
        AuthService.LoginResult result = authService.login(request);
        return Map.of("token", result.token(), "user", UserResponse.of(result.user()));
    }

    @GetMapping("/me")
    public Map<String, Object> me(@RequestHeader(value = HttpHeaders.AUTHORIZATION, required = false) String header) {
        return Map.of("user", UserResponse.of(authService.requireUser(BearerToken.from(header))));
    }

    @PostMapping("/logout")
    public Map<String, String> logout(
            @RequestHeader(value = HttpHeaders.AUTHORIZATION, required = false) String header) {
        authService.logout(BearerToken.from(header));
        return Map.of("status", "signed-out");
    }
}